move files to map directory
