+++ components/map/layer-manager/component.js
