+++ components/map/layer-manager/index.js
