+++ components/map/layer-manager/utils.js
