+++ components/map/legend/analysis/gain/component.js
