+++ components/map/legend/analysis/gain/index.js
