+++ components/map/legend/analysis/index.js
