+++ components/map/legend/analysis/integrated-alerts/component.js
