+++ components/map/legend/analysis/integrated-alerts/index.js
