+++ components/map/legend/analysis/loss/component.js
