+++ components/map/legend/analysis/loss/index.js
