+++ components/map/legend/component.js
