+++ components/map/legend/index.js
