+++ components/map/legend/templates/alerts/component.js
