+++ components/map/legend/templates/alerts/index.js
