+++ components/map/legend/templates/index.js
