+++ components/ui/sawmill-modal.js
@@ -22,8 +22,8 @@ import Field from 'components/form/Field';
 import Input from 'components/form/Input';
 import Checkbox from 'components/form/Checkbox';
 import Spinner from 'components/ui/spinner';
-import Map from 'components/map-new';
-import LayerManager from 'components/map-new/layer-manager';
+import Map from 'components/map';
+import LayerManager from 'components/map/layer-manager';
 import MapControls from 'components/map/map-controls';
 import ZoomControl from 'components/map/controls/zoom-control';
 import LocationSearch from 'components/map/location-search';
