+++ pages/operators/index.js
@@ -33,11 +33,11 @@ import modal from 'services/modal';
 // Components
 import Layout from 'components/layout/layout';
 import Sidebar from 'components/ui/sidebar';
-import Map from 'components/map-new';
-import LayerManager from 'components/map-new/layer-manager';
-import Popup from 'components/map-new/popup';
-import Legend from 'components/map-new/legend';
-import FAAttributions from 'components/map-new/fa-attributions';
+import Map from 'components/map';
+import LayerManager from 'components/map/layer-manager';
+import Popup from 'components/map/popup';
+import Legend from 'components/map/legend';
+import FAAttributions from 'components/map/fa-attributions';
 
 import MapControls from 'components/map/map-controls';
 import ZoomControl from 'components/map/controls/zoom-control';
